+++ internal/download/ugoira_rust_link_linux_arm64.go
@@ -3,6 +3,6 @@
 package download
 
 /*
-#cgo LDFLAGS: ${SRCDIR}/ugoira_rs/staticlib/aarch64-unknown-linux-gnu/libugoira_rs.a
+#cgo LDFLAGS: ${SRCDIR}/ugoira_rs/staticlib/aarch64-unknown-linux-gnu/libugoira_rs.a -lm
 */
 import "C"